[BE] refactor: router 로직 단일화
미들웨어를 하나로 통합하면서 public router를
그냥 router로 내림
