+++ backend/routes/dj-stations/index.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import djStationController from "../../../controllers/public/dj-stations";
+import djStationController from "../../controllers/public/dj-stations";
 
 const router = Router();
 router.get("/", djStationController.getAll);
