+++ backend/routes/genres/index.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import genreController from "../../../controllers/public/genres";
+import genreController from "../../controllers/public/genres";
 
 const router = Router();
 router.get("/", genreController.getAll);
